requirements.txt
asgiref==3.9.1
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.2
cryptography==45.0.5
Django==5.2.4
django-appconf==1.1.0
django-cryptography==1.1
django-fernet-fields==0.6
django-otp==1.6.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
pyAesCrypt==6.1.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
qrcode==8.2
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
Werkzeug==3.1.3
